+++ resolwe_bio/docker_images/legacy/packages-manual/samtools.sh
@@ -1,0 +1,21 @@
+#!/bin/bash -e
+
+. /var/cache/build/packages-manual/common.sh
+
+download_and_verify \
+    samtools \
+    samtools \
+    1.3.1 \
+    6c3d74355e9cf2d9b2e1460273285d154107659efa36a155704b1e4358b7d67e \
+    https://github.com/samtools/samtools/releases/download/\${version}/samtools-\${version}.tar.bz2 \
+    samtools-\${version}
+
+./configure --without-curses
+make
+
+shopt -s extglob
+rm -rf !(samtools)
+
+add_binary_path \
+    samtools \
+    samtools